complete split sentene in smaller chunks
