+++ src/components/practiceTool/PracticeQuestion.tsx
@@ -437,23 +437,6 @@ export const PracticeQuestion = ({
 
       {question && !practiceIsCompleted && (
         <>
-          {/* Assistant */}
-          {enabledAssistant && (
-            <Box
-              sx={{
-                position: "fixed",
-                bottom: "50px",
-                right: "50px",
-              }}
-            >
-              <img
-                src={`/assistant.svg`}
-                style={{
-                  width: "70px",
-                }}
-              />
-            </Box>
-          )}
           {/* options */}
           <Stack spacing={"8px"} sx={{ position: "absolute", right: "12px", top: "8px" }}>
             <IconButton
