+++ src/components/practiceTool/PracticeTool.tsx
@@ -9,7 +9,7 @@ import React, {
   useImperativeHandle,
   useState,
 } from "react";
-import { TVoiceAssistantRef } from "src/nodeBookTypes";
+import { VoiceAssistant } from "src/nodeBookTypes";
 import { ISemester } from "src/types/ICourse";
 import { ISemesterStudentVoteStat } from "src/types/ICourse";
 
@@ -25,12 +25,15 @@ import { PracticeQuestion } from "./PracticeQuestion";
 import { UserStatus } from "./UserStatus";
 
 type PracticeToolProps = {
-  setVoiceAssistant: Dispatch<SetStateAction<TVoiceAssistantRef | null>>;
+  voiceAssistant: VoiceAssistant | null;
+  setVoiceAssistant: Dispatch<SetStateAction<VoiceAssistant | null>>;
   user: User;
   root?: string;
   currentSemester: CourseTag;
   onClose: () => void;
   openNodeHandler: (nodeId: string) => void;
+  enabledAssistant: boolean;
+  setEnabledAssistant: Dispatch<SetStateAction<boolean>>;
 };
 
 export type PracticeInfo = {
@@ -57,195 +60,202 @@ export type PracticeToolRef = {
   getQuestionParents: () => string[];
 };
 
-const PracticeTool = forwardRef<PracticeToolRef, PracticeToolProps>(
-  ({ setVoiceAssistant, user, currentSemester, openNodeHandler, onClose, root }, ref) => {
-    console.log({ currentSemester });
-    const db = getFirestore();
-    const [startPractice, setStartPractice] = useState(false);
-    const [questionData, setQuestionData] = useState<{ question: SimpleQuestionNode; flashcardId: string } | null>(
-      null
-    );
-    const [practiceIsCompleted, setPracticeIsCompleted] = useState(false);
-    const [practiceInfo, setPracticeInfo] = useState<PracticeInfo>(DEFAULT_PRACTICE_INFO);
-    const [semesterConfig, setSemesterConfig] = useState<ISemester | null>(null);
-    const [submitAnswer, setSubmitAnswer] = useState(false);
-    const [selectedAnswers, setSelectedAnswers] = useState<boolean[]>([]);
-    const [enabledAssistant, setEnabledAssistant] = useState(false);
-    const onRunPracticeTool = useCallback(() => {
-      (start: boolean) => {
-        if (!practiceInfo) return;
+const PracticeTool = forwardRef<PracticeToolRef, PracticeToolProps>((props, ref) => {
+  const {
+    voiceAssistant,
+    setVoiceAssistant,
+    user,
+    currentSemester,
+    openNodeHandler,
+    onClose,
+    root,
+    // enabledAssistant,
+    // setEnabledAssistant,
+  } = props;
+  console.log({ currentSemester });
+  const db = getFirestore();
+  const [startPractice, setStartPractice] = useState(false);
+  const [questionData, setQuestionData] = useState<{ question: SimpleQuestionNode; flashcardId: string } | null>(null);
+  const [practiceIsCompleted, setPracticeIsCompleted] = useState(false);
+  const [practiceInfo, setPracticeInfo] = useState<PracticeInfo>(DEFAULT_PRACTICE_INFO);
+  const [semesterConfig, setSemesterConfig] = useState<ISemester | null>(null);
+  const [submitAnswer, setSubmitAnswer] = useState(false);
+  const [selectedAnswers, setSelectedAnswers] = useState<boolean[]>([]);
+  const [enabledAssistant, setEnabledAssistant] = useState(Boolean(voiceAssistant));
+  const onRunPracticeTool = useCallback(() => {
+    (start: boolean) => {
+      if (!practiceInfo) return;
 
-        setStartPractice(start);
+      setStartPractice(start);
+    };
+  }, [practiceInfo]);
+
+  const onSubmitAnswer = useCallback(
+    async (answers: boolean[]) => {
+      if (!questionData) return;
+
+      console.log("onSubmitAnswer", answers);
+      setSubmitAnswer(true);
+
+      const payload: ICheckAnswerRequestParams = {
+        answers,
+        flashcardId: questionData.flashcardId,
+        nodeId: questionData.question.id,
+        postpone: false,
       };
-    }, [practiceInfo]);
+      await Post("/checkAnswer", payload).then(() => {});
+    },
+    [questionData]
+  );
 
-    const onSubmitAnswer = useCallback(
-      async (answers: boolean[]) => {
-        if (!questionData) return;
+  const getPracticeQuestion = useCallback(async () => {
+    const res: any = await Post("/practice", { tagId: currentSemester.tagId });
+    console.log("practice:res", { res });
+    if (res?.done) return setPracticeIsCompleted(true);
 
-        console.log("onSubmitAnswer", answers);
-        setSubmitAnswer(true);
+    const question = res.question as SimpleQuestionNode;
+    setSubmitAnswer(false);
+    setSelectedAnswers(new Array(question.choices.length).fill(false));
+    setQuestionData(res);
+  }, [currentSemester.tagId]);
 
-        const payload: ICheckAnswerRequestParams = {
-          answers,
-          flashcardId: questionData.flashcardId,
-          nodeId: questionData.question.id,
-          postpone: false,
-        };
-        await Post("/checkAnswer", payload).then(() => {});
-      },
-      [questionData]
+  useImperativeHandle(ref, () => ({
+    onRunPracticeTool,
+    onSubmitAnswer,
+    onSelectAnswers: answers => setSelectedAnswers(answers),
+    nextQuestion: getPracticeQuestion,
+    getQuestionParents: () => questionData?.question.parents ?? [],
+  }));
+
+  const onViewNodeOnNodeBook = (nodeId: string) => {
+    openNodeHandler(nodeId);
+    onClose();
+  };
+
+  // this is executed the first time we get selected a semester
+  useEffect(() => {
+    if (!startPractice) return;
+    getPracticeQuestion();
+  }, [getPracticeQuestion, startPractice]);
+
+  useEffect(() => {
+    const getSemesterConfig = async () => {
+      const semester = await getSemesterById(db, currentSemester.tagId);
+      if (!semester) return;
+
+      setSemesterConfig(semester);
+    };
+    getSemesterConfig();
+  }, [currentSemester.tagId, db, user.uname]);
+
+  useEffect(() => {
+    console.log({ semesterConfig });
+    if (!semesterConfig) return;
+
+    const q = query(
+      collection(db, "semesterStudentVoteStats"),
+      where("uname", "==", user.uname),
+      where("tagId", "==", currentSemester.tagId)
     );
+    const unsub = onSnapshot(q, snapshot => {
+      if (snapshot.empty) return;
 
-    const getPracticeQuestion = useCallback(async () => {
-      const res: any = await Post("/practice", { tagId: currentSemester.tagId });
-      console.log("practice:res", { res });
-      if (res?.done) return setPracticeIsCompleted(true);
+      const docChanges = snapshot.docChanges();
+      for (const docChange of docChanges) {
+        const semesterStudentVoteStat = docChange.doc.data() as ISemesterStudentVoteStat;
+        console.log({ semesterStudentVoteStat });
+        const currentDateYYMMDD = getDateYYMMDDWithHyphens();
+        console.log({ currentDateYYMMDD });
+        const currentDayStats = semesterStudentVoteStat.days.find(cur => cur.day === currentDateYYMMDD);
+        console.log({ currentDayStats });
+        // if (!currentDayStats) return;
+        //
+        const totalQuestions = semesterConfig.dailyPractice.numQuestionsPerDay;
+        const questionsLeft = totalQuestions - (currentDayStats?.correctPractices ?? 0);
+        console.log({ questionsLeft });
+        // setPracticeInfo(prev => ({ ...prev, questionsLeft }));
 
-      const question = res.question as SimpleQuestionNode;
-      setSubmitAnswer(false);
-      setSelectedAnswers(new Array(question.choices.length).fill(false));
-      setQuestionData(res);
-    }, [currentSemester.tagId]);
+        const completedDays = differentBetweenDays(new Date(), semesterConfig.startDate.toDate());
+        const totalDays = differentBetweenDays(semesterConfig.endDate.toDate(), semesterConfig.startDate.toDate());
+        const remainingDays = totalDays - completedDays;
+        setPracticeInfo(prev => ({
+          ...prev,
+          questionsLeft,
+          totalQuestions,
+          completedDays,
+          totalDays,
+          remainingDays,
+        }));
+      }
+    });
+    return () => {
+      if (unsub) unsub();
+    };
+  }, [currentSemester.tagId, db, semesterConfig, user.uname]);
 
-    useImperativeHandle(ref, () => ({
-      onRunPracticeTool,
-      onSubmitAnswer,
-      onSelectAnswers: answers => setSelectedAnswers(answers),
-      nextQuestion: getPracticeQuestion,
-      getQuestionParents: () => questionData?.question.parents ?? [],
-    }));
+  useEffect(() => {
+    console.log({ practiceInfo });
+    if (!practiceInfo) return;
+    if (!semesterConfig) return;
+    if (!root) return;
 
-    const onViewNodeOnNodeBook = (nodeId: string) => {
-      openNodeHandler(nodeId);
-      onClose();
+    setStartPractice(true);
+  }, [practiceInfo, root, semesterConfig]);
+
+  useEffect(() => {
+    const detectAssistantEnable = () => {
+      console.log("detectAssistantEnable", { enabledAssistant, questionData });
+      if (!enabledAssistant) return setVoiceAssistant(null);
+      if (!questionData) return;
+
+      const choiceMessage = questionData.question.choices.map(cur => cur.choice.replace(".", ",")).join(". ");
+      setVoiceAssistant({
+        state: "NARRATE",
+        listenType: "ANSWERING",
+        message: `${questionData.question.title}. ${choiceMessage}`,
+        answers: questionData.question.choices,
+        selectedAnswer: "",
+        date: "",
+        tagId: currentSemester.tagId,
+      });
     };
 
-    // this is executed the first time we get selected a semester
-    useEffect(() => {
-      if (!startPractice) return;
-      getPracticeQuestion();
-    }, [getPracticeQuestion, startPractice]);
+    detectAssistantEnable();
+  }, [currentSemester.tagId, enabledAssistant, questionData, setVoiceAssistant]);
 
-    useEffect(() => {
-      const getSemesterConfig = async () => {
-        const semester = await getSemesterById(db, currentSemester.tagId);
-        if (!semester) return;
-
-        setSemesterConfig(semester);
-      };
-      getSemesterConfig();
-    }, [currentSemester.tagId, db, user.uname]);
-
-    useEffect(() => {
-      console.log({ semesterConfig });
-      if (!semesterConfig) return;
-
-      const q = query(
-        collection(db, "semesterStudentVoteStats"),
-        where("uname", "==", user.uname),
-        where("tagId", "==", currentSemester.tagId)
-      );
-      const unsub = onSnapshot(q, snapshot => {
-        if (snapshot.empty) return;
-
-        const docChanges = snapshot.docChanges();
-        for (const docChange of docChanges) {
-          const semesterStudentVoteStat = docChange.doc.data() as ISemesterStudentVoteStat;
-          console.log({ semesterStudentVoteStat });
-          const currentDateYYMMDD = getDateYYMMDDWithHyphens();
-          console.log({ currentDateYYMMDD });
-          const currentDayStats = semesterStudentVoteStat.days.find(cur => cur.day === currentDateYYMMDD);
-          console.log({ currentDayStats });
-          // if (!currentDayStats) return;
-          //
-          const totalQuestions = semesterConfig.dailyPractice.numQuestionsPerDay;
-          const questionsLeft = totalQuestions - (currentDayStats?.correctPractices ?? 0);
-          console.log({ questionsLeft });
-          // setPracticeInfo(prev => ({ ...prev, questionsLeft }));
-
-          const completedDays = differentBetweenDays(new Date(), semesterConfig.startDate.toDate());
-          const totalDays = differentBetweenDays(semesterConfig.endDate.toDate(), semesterConfig.startDate.toDate());
-          const remainingDays = totalDays - completedDays;
-          setPracticeInfo(prev => ({
-            ...prev,
-            questionsLeft,
-            totalQuestions,
-            completedDays,
-            totalDays,
-            remainingDays,
-          }));
-        }
-      });
-      return () => {
-        if (unsub) unsub();
-      };
-    }, [currentSemester.tagId, db, semesterConfig, user.uname]);
-
-    useEffect(() => {
-      console.log({ practiceInfo });
-      if (!practiceInfo) return;
-      if (!semesterConfig) return;
-      if (!root) return;
-
-      setStartPractice(true);
-    }, [practiceInfo, root, semesterConfig]);
-
-    useEffect(() => {
-      const detectAssistantEnable = () => {
-        if (!enabledAssistant) return setVoiceAssistant(null);
-        if (!questionData) return;
-
-        const choiceMessage = questionData.question.choices.map(cur => cur.choice.replace(".", ",")).join(". ");
-        setVoiceAssistant({
-          listen: false,
-          listenType: "ANSWERING",
-          message: `${questionData.question.title}. ${choiceMessage}`,
-          narrate: true,
-          answers: questionData.question.choices,
-          selectedAnswer: "",
-          date: "",
-          tagId: currentSemester.tagId,
-        });
-      };
-
-      detectAssistantEnable();
-    }, [currentSemester.tagId, enabledAssistant, questionData, setVoiceAssistant]);
-
-    return startPractice ? (
-      <Box
-        sx={{
-          position: "absolute",
-          inset: "0px",
-          background: theme =>
-            theme.palette.mode === "dark" ? theme.palette.common.notebookG900 : theme.palette.common.notebookBl1,
-          zIndex: 1,
-          overflowY: "auto",
-          overflowX: "hidden",
-        }}
-      >
-        <PracticeQuestion
-          question={questionData?.question ?? null}
-          practiceIsCompleted={practiceIsCompleted}
-          onClose={onClose}
-          leaderboard={<Leaderboard semesterId={currentSemester.tagId} />}
-          userStatus={<UserStatus semesterId={currentSemester.tagId} user={user} />}
-          onViewNodeOnNodeBook={onViewNodeOnNodeBook}
-          onGetNextQuestion={getPracticeQuestion}
-          onSaveAnswer={onSubmitAnswer}
-          practiceInfo={practiceInfo}
-          submitAnswer={submitAnswer}
-          setSubmitAnswer={setSubmitAnswer}
-          selectedAnswers={selectedAnswers}
-          setSelectedAnswers={setSelectedAnswers}
-          enabledAssistant={enabledAssistant}
-          setEnabledAssistant={setEnabledAssistant}
-        />
-      </Box>
-    ) : (
-      <CourseDetail user={user} currentSemester={currentSemester} onStartPractice={() => setStartPractice(true)} />
-    );
-  }
-);
+  return startPractice ? (
+    <Box
+      sx={{
+        position: "absolute",
+        inset: "0px",
+        background: theme =>
+          theme.palette.mode === "dark" ? theme.palette.common.notebookG900 : theme.palette.common.notebookBl1,
+        zIndex: 1,
+        overflowY: "auto",
+        overflowX: "hidden",
+      }}
+    >
+      <PracticeQuestion
+        question={questionData?.question ?? null}
+        practiceIsCompleted={practiceIsCompleted}
+        onClose={onClose}
+        leaderboard={<Leaderboard semesterId={currentSemester.tagId} />}
+        userStatus={<UserStatus semesterId={currentSemester.tagId} user={user} />}
+        onViewNodeOnNodeBook={onViewNodeOnNodeBook}
+        onGetNextQuestion={getPracticeQuestion}
+        onSaveAnswer={onSubmitAnswer}
+        practiceInfo={practiceInfo}
+        submitAnswer={submitAnswer}
+        setSubmitAnswer={setSubmitAnswer}
+        selectedAnswers={selectedAnswers}
+        setSelectedAnswers={setSelectedAnswers}
+        enabledAssistant={Boolean(voiceAssistant)}
+        setEnabledAssistant={setEnabledAssistant}
+      />
+    </Box>
+  ) : (
+    <CourseDetail user={user} currentSemester={currentSemester} onStartPractice={() => setStartPractice(true)} />
+  );
+});
 PracticeTool.displayName = "PracticeTool";
 export default PracticeTool;
