+++ src/lib/utils/string.utils.ts
@@ -20,3 +20,24 @@ export const ellipsisString = (text: string, length: number) => {
 export const getTextSplittedByCharacter = (text: string, character: string): string => {
   return Array.from(text).join(character);
 };
+
+export function splitSentenceIntoChunks(sentence: string, maxCharacters = 100): string[] {
+  const words: string[] = sentence.split(" ");
+  const chunks: string[] = [];
+  let currentChunk: string = "";
+
+  for (const word of words) {
+    if (currentChunk.length + word.length <= maxCharacters) {
+      currentChunk += (currentChunk.length > 0 ? " " : "") + word;
+    } else {
+      chunks.push(currentChunk);
+      currentChunk = word;
+    }
+  }
+
+  if (currentChunk.length > 0) {
+    chunks.push(currentChunk);
+  }
+
+  return chunks;
+}
