+++ src/pages/notebook.tsx
@@ -64,13 +64,13 @@ import { MemoizedUserInfoSidebar } from "@/components/map/Sidebar/SidebarV2/User
 import { MemoizedUserSettingsSidebar } from "@/components/map/Sidebar/SidebarV2/UserSettigsSidebar";
 import { useAuth } from "@/context/AuthContext";
 import useEventListener from "@/hooks/useEventListener";
-import usePrevious from "@/hooks/usePrevious";
+// import usePrevious from "@/hooks/usePrevious";
 import { useTagsTreeView } from "@/hooks/useTagsTreeView";
 import { DESIGN_SYSTEM_COLORS } from "@/lib/theme/colors";
 
 import LoadingImg from "../../public/animated-icon-1cademy.gif";
-import { getNode } from "../client/serveless/nodes.serveless";
 import { TooltipTutorial } from "../components/interactiveTutorial/Tutorial";
+import { Assistant } from "../components/map/Assistant";
 // import nodesData from "../../testUtils/mockCollections/nodes.data";
 // import { Tutorial } from "../components/interactiveTutorial/Tutorial";
 import { MemoizedClustersList } from "../components/map/ClustersList";
@@ -94,19 +94,10 @@ import {
 import { useMemoizedCallback } from "../hooks/useMemoizedCallback";
 import { useWindowSize } from "../hooks/useWindowSize";
 import { useWorkerQueue } from "../hooks/useWorkerQueue";
-import { KnowledgeChoice, NodeChanges, ReputationSignal } from "../knowledgeTypes";
+import { NodeChanges, ReputationSignal } from "../knowledgeTypes";
 import { idToken, retrieveAuthenticatedUser } from "../lib/firestoreClient/auth";
 import { Post, postWithToken } from "../lib/mapApi";
-import { getAnswersLettersOptions } from "../lib/utils/assistant.utils";
-import {
-  ANSWERING_ERROR,
-  ASSISTANT_NEGATIVE_SENTENCES,
-  ASSISTANT_POSITIVE_SENTENCES,
-  CONFIRM_ERROR,
-  NEXT_ACTION_ERROR,
-  NO_USER_IMAGE,
-  QUESTION_OPTIONS,
-} from "../lib/utils/constants";
+import { NO_USER_IMAGE, ZINDEX } from "../lib/utils/constants";
 import { createGraph, dagreUtils } from "../lib/utils/dagre.util";
 import { devLog } from "../lib/utils/develop.util";
 import { getTypedCollections } from "../lib/utils/getTypedCollections";
@@ -140,8 +131,6 @@ import {
   getUserNodeChanges,
   mergeAllNodes,
 } from "../lib/utils/nodesSyncronization.utils";
-import { getValidABCDOptions, newRecognition } from "../lib/utils/speechRecognitions.utils";
-import { getTextSplittedByCharacter } from "../lib/utils/string.utils";
 import { getGroupTutorials, LivelinessBar } from "../lib/utils/tutorials/grouptutorials";
 import { gtmEvent, imageLoaded, isValidHttpUrl } from "../lib/utils/utils";
 import {
@@ -149,23 +138,20 @@ import {
   EdgesData,
   FullNodeData,
   FullNodesData,
-  Node,
   OpenPart,
   // NodeTutorialState,
   TNodeBookState,
   TNodeUpdates,
   TutorialTypeKeys,
-  TVoiceAssistantListenType,
-  TVoiceAssistantRef,
   // TutorialType,
   UserNodes,
   UserNodesData,
   UserTutorial,
   UserTutorials,
+  VoiceAssistant,
 } from "../nodeBookTypes";
 import { NodeType, Notebook, NotebookDocument, SimpleNode2 } from "../types";
-import { doNeedToDeleteNode, getNodeTypesFromNode, isVersionApproved, narrateLargeTexts } from "../utils/helpers";
-import { nodeToNarration } from "../utils/node.utils";
+import { doNeedToDeleteNode, getNodeTypesFromNode, isVersionApproved } from "../utils/helpers";
 
 type NotebookProps = {};
 
@@ -382,6 +368,7 @@ const Notebook = ({}: NotebookProps) => {
   const [notebooks, setNotebooks] = useState<Notebook[]>([]);
   const [selectedNotebookId, setSelectedNotebookId] = useState("");
   const selectedPreviousNotebookIdRef = useRef("");
+  const [enabledAssistant, setEnabledAssistant] = useState(false);
 
   const onChangeTagOfNotebookById = (notebookId: string, data: { defaultTagId: string; defaultTagName: string }) => {
     setNotebooks(prev => {
@@ -669,8 +656,8 @@ const Notebook = ({}: NotebookProps) => {
   // const [voiceAssistantUpdates, setVoiceAssistantUpdates] = useState({
   //   updated: new Date(),
   // });
-  const [voiceAssistant, setVoiceAssistant] = useState<TVoiceAssistantRef | null>(null);
-  const prevVoiceAssistant = usePrevious(voiceAssistant);
+  const [voiceAssistant, setVoiceAssistant] = useState<VoiceAssistant | null>(null);
+  // const prevVoiceAssistant = usePrevious(voiceAssistant);
 
   const assistantRef = useRef<DashboardWrapperRef | null>(null);
 
@@ -6185,348 +6172,346 @@ const Notebook = ({}: NotebookProps) => {
     duplicateNotebookFromParams();
   }, [db, onChangeNotebook, openNodesOnNotebook, router.query.nb, user]);
 
-  // assistant will narrate and then will listen
-  useEffect(() => {
-    const assistantActions = async () => {
-      if (prevVoiceAssistant && !voiceAssistant) {
-        window.speechSynthesis.cancel();
-        const message = "Assistant stopped";
-        await narrateLargeTexts(message);
-      }
+  // // assistant will narrate and then will listen
+  // useEffect(() => {
+  //   const assistantActions = async () => {
+  //     if (prevVoiceAssistant && !voiceAssistant) {
+  //       window.speechSynthesis.cancel();
+  //       const message = "Assistant stopped";
+  //       await narrateLargeTexts(message);
+  //     }
 
-      if (!voiceAssistant) {
-        return;
-      }
+  //     if (!voiceAssistant) {
+  //       return;
+  //     }
 
-      const recognition = newRecognition(voiceAssistant.listenType as TVoiceAssistantListenType);
-      if (!recognition) return console.error("This browser does't support speech recognition");
+  //     const recognition = newRecognition(voiceAssistant.listenType as VoiceAssistantType);
+  //     if (!recognition) return console.error("This browser does't support speech recognition");
 
-      console.log("👉 assistantActions:start:narrate", voiceAssistant);
-      await narrateLargeTexts(voiceAssistant.message);
+  //     console.log("👉 assistantActions:start:narrate", voiceAssistant);
+  //     await narrateLargeTexts(voiceAssistant.message);
 
-      console.log("assistant:11", voiceAssistant.listenType);
-      if (!voiceAssistant.listenType) return;
-      console.log("assistant:22", voiceAssistant.listenType);
-      const NEXT_ACTION = "*";
-      const OPEN_NOTEBOOK = ".";
-      const REPEAT_QUESTION = "?";
-      const OPEN_PRACTICE = "#";
+  //     console.log("assistant:11", voiceAssistant.listenType);
+  //     if (!voiceAssistant.listenType) return;
+  //     console.log("assistant:22", voiceAssistant.listenType);
+  //     const NEXT_ACTION = "*";
+  //     const OPEN_NOTEBOOK = ".";
+  //     const REPEAT_QUESTION = "?";
+  //     const OPEN_PRACTICE = "#";
 
-      const MapSentences: { [key: string]: string } = {
-        "repeat question": REPEAT_QUESTION,
-        "open notebook": OPEN_NOTEBOOK,
-        "continue practicing": OPEN_PRACTICE,
-      };
-      const MapWords: { [key: string]: string } = {
-        hey: "a",
-        be: "b",
-        ve: "b",
-        me: "b",
-        ce: "c",
-        see: "c",
-        se: "c",
-        de: "d",
-        dee: "d",
-        guess: "d",
-        he: "e",
-        yes: "y",
-        correct: "y",
-        "repeat question": REPEAT_QUESTION,
-        next: NEXT_ACTION,
-        nikes: NEXT_ACTION,
-        "open notebook": OPEN_NOTEBOOK,
-        "continue practicing": OPEN_PRACTICE,
-      };
-      recognition.start();
-      recognition.onresult = async (event: any) => {
-        const transcript: string = event.results?.[0]?.[0]?.transcript || "";
-        console.log("----> result", { transcript });
+  //     const MapSentences: { [key: string]: string } = {
+  //       "repeat question": REPEAT_QUESTION,
+  //       "open notebook": OPEN_NOTEBOOK,
+  //       "continue practicing": OPEN_PRACTICE,
+  //     };
+  //     const MapWords: { [key: string]: string } = {
+  //       hey: "a",
+  //       be: "b",
+  //       ve: "b",
+  //       me: "b",
+  //       ce: "c",
+  //       see: "c",
+  //       se: "c",
+  //       de: "d",
+  //       dee: "d",
+  //       guess: "d",
+  //       he: "e",
+  //       yes: "y",
+  //       correct: "y",
+  //       "repeat question": REPEAT_QUESTION,
+  //       next: NEXT_ACTION,
+  //       nikes: NEXT_ACTION,
+  //       "open notebook": OPEN_NOTEBOOK,
+  //       "continue practicing": OPEN_PRACTICE,
+  //     };
+  //     recognition.start();
+  //     recognition.onresult = async (event: any) => {
+  //       const transcript: string = event.results?.[0]?.[0]?.transcript || "";
+  //       console.log("----> result", { transcript });
 
-        // here call directly important commands
+  //       // here call directly important commands
 
-        if ("repeat question" === transcript.toLowerCase()) {
-          setVoiceAssistant({
-            ...voiceAssistant,
-            listen: false,
-            listenType: "ANSWERING",
-            narrate: true,
-            selectedAnswer: "",
-          });
-          return;
-        }
-        if ("stop" === transcript.toLowerCase()) {
-          const message = "Assistant stopped";
-          await narrateLargeTexts(message);
-          setVoiceAssistant(null);
-          return;
-        }
+  //       if ("repeat question" === transcript.toLowerCase()) {
+  //         setVoiceAssistant({
+  //           ...voiceAssistant,
+  //           listen: false,
+  //           listenType: "ANSWERING",
+  //           narrate: true,
+  //           selectedAnswer: "",
+  //         });
+  //         return;
+  //       }
+  //       if ("stop" === transcript.toLowerCase()) {
+  //         const message = "Assistant stopped";
+  //         await narrateLargeTexts(message);
+  //         setVoiceAssistant(null);
+  //         return;
+  //       }
 
-        // here process the transcript to correct most possible transcript value
-        let possibleTranscript: string | null = null;
-        if (voiceAssistant.listenType === "ANSWERING")
-          possibleTranscript = getValidABCDOptions(transcript.toLowerCase()); // if is answering and is valid, we use directly
+  //       // here process the transcript to correct most possible transcript value
+  //       let possibleTranscript: string | null = null;
+  //       if (voiceAssistant.listenType === "ANSWERING")
+  //         possibleTranscript = getValidABCDOptions(transcript.toLowerCase()); // if is answering and is valid, we use directly
 
-        const transcriptProcessed =
-          possibleTranscript ??
-          MapSentences[transcript.toLowerCase()] ??
-          transcript
-            .toLowerCase()
-            .split(" ")
-            .map(cur => (cur.length === 1 ? cur : MapWords[cur] ?? ""))
-            .filter(cur => cur.length === 1)
-            .join("");
-        console.log("--->", { transcriptProcessed });
-        // actions to interrupt normal flow
+  //       const transcriptProcessed =
+  //         possibleTranscript ??
+  //         MapSentences[transcript.toLowerCase()] ??
+  //         transcript
+  //           .toLowerCase()
+  //           .split(" ")
+  //           .map(cur => (cur.length === 1 ? cur : MapWords[cur] ?? ""))
+  //           .filter(cur => cur.length === 1)
+  //           .join("");
+  //       console.log("--->", { transcriptProcessed });
+  //       // actions to interrupt normal flow
 
-        // INFO: pause and resumen is not possible because will work like
-        //       and infinity loop waiting on silence until user tell resumen
-        //       that need to executed into useEffect so render will be called in every iteration
+  //       // INFO: pause and resumen is not possible because will work like
+  //       //       and infinity loop waiting on silence until user tell resumen
+  //       //       that need to executed into useEffect so render will be called in every iteration
 
-        // if ("pause" === transcript.toLowerCase()) {
-        //   const message = "Assistant paused, please tell me resume to continue";
-        //   await narrateLargeTexts(message);
-        //   setVoiceAssistant(VOICE_ASSISTANT_DEFAULT);
-        //   return;
-        // }
+  //       // if ("pause" === transcript.toLowerCase()) {
+  //       //   const message = "Assistant paused, please tell me resume to continue";
+  //       //   await narrateLargeTexts(message);
+  //       //   setVoiceAssistant(VOICE_ASSISTANT_DEFAULT);
+  //       //   return;
+  //       // }
 
-        // if ("resume" === transcript.toLowerCase()) {
-        //   const message = "I am resuming my previous task";
-        //   await narrateLargeTexts(message);
-        //   setVoiceAssistant(prev => ({ ...prev, date: "from-resume-assistant" }));
-        //   return;
-        // }
+  //       // if ("resume" === transcript.toLowerCase()) {
+  //       //   const message = "I am resuming my previous task";
+  //       //   await narrateLargeTexts(message);
+  //       //   setVoiceAssistant(prev => ({ ...prev, date: "from-resume-assistant" }));
+  //       //   return;
+  //       // }
 
-        // if ("help" === transcript.toLowerCase()) {
-        //   const message = "Assistant Paused, Please tell me resume";
-        //   await narrateLargeTexts(message);
-        //   setVoiceAssistant(VOICE_ASSISTANT_DEFAULT);
-        //   return;
-        // }
+  //       // if ("help" === transcript.toLowerCase()) {
+  //       //   const message = "Assistant Paused, Please tell me resume";
+  //       //   await narrateLargeTexts(message);
+  //       //   setVoiceAssistant(VOICE_ASSISTANT_DEFAULT);
+  //       //   return;
+  //       // }
 
-        // no valid answers will ask again the same question
-        if (!transcriptProcessed && voiceAssistant.listenType !== "CONFIRM") {
-          console.log("No transcription", voiceAssistant.listenType);
-          let message = "Sorry, I didn't get your choices.";
-          // if (voiceAssistant.listenType === "CONFIRM") {
-          //   message += CONFIRM_ERROR;
-          // }
-          if (voiceAssistant.listenType === "ANSWERING") {
-            message += ANSWERING_ERROR;
-          }
-          if (voiceAssistant.listenType === "NEXT_ACTION") {
-            message += NEXT_ACTION_ERROR;
-          }
+  //       // no valid answers will ask again the same question
+  //       if (!transcriptProcessed && voiceAssistant.listenType !== "CONFIRM") {
+  //         console.log("No transcription", voiceAssistant.listenType);
+  //         let message = "Sorry, I didn't get your choices.";
+  //         // if (voiceAssistant.listenType === "CONFIRM") {
+  //         //   message += CONFIRM_ERROR;
+  //         // }
+  //         if (voiceAssistant.listenType === "ANSWERING") {
+  //           message += ANSWERING_ERROR;
+  //         }
+  //         if (voiceAssistant.listenType === "NEXT_ACTION") {
+  //           message += NEXT_ACTION_ERROR;
+  //         }
 
-          await narrateLargeTexts(message);
-          setVoiceAssistant({
-            ...voiceAssistant,
-            date: "from-empty transcript",
-            message,
-            // message: voiceAssistant.listenType === "NEXT_ACTION" ? "" : voiceAssistant.message,
-          });
-          return;
-        }
+  //         await narrateLargeTexts(message);
+  //         setVoiceAssistant({
+  //           ...voiceAssistant,
+  //           date: "from-empty transcript",
+  //           message,
+  //           // message: voiceAssistant.listenType === "NEXT_ACTION" ? "" : voiceAssistant.message,
+  //         });
+  //         return;
+  //       }
 
-        // actions according the flow
-        if (voiceAssistant.listenType === "ANSWERING") {
-          if (!assistantRef.current) return;
-          // transcriptProcessed:"bc"
-          // possibleOptions: "abcd"
-          const possibleOptions = QUESTION_OPTIONS.slice(0, voiceAssistant.answers.length);
-          const answerIsValid = Array.from(transcriptProcessed).reduce(
-            (acu, cur) => acu && possibleOptions.includes(cur),
-            true
-          );
-          console.log("assistantActions:ANSWERING", { possibleOptions, transcriptProcessed, answerIsValid });
-          if (!answerIsValid) {
-            const message = ANSWERING_ERROR;
-            setVoiceAssistant({
-              ...voiceAssistant,
-              listen: false,
-              listenType: "ANSWERING",
-              message,
-              narrate: true,
-              selectedAnswer: "",
-            });
-            return;
-          }
+  //       // actions according the flow
+  //       if (voiceAssistant.listenType === "ANSWERING") {
+  //         if (!assistantRef.current) return;
+  //         // transcriptProcessed:"bc"
+  //         // possibleOptions: "abcd"
+  //         const possibleOptions = QUESTION_OPTIONS.slice(0, voiceAssistant.answers.length);
+  //         const answerIsValid = Array.from(transcriptProcessed).reduce(
+  //           (acu, cur) => acu && possibleOptions.includes(cur),
+  //           true
+  //         );
+  //         console.log("assistantActions:ANSWERING", { possibleOptions, transcriptProcessed, answerIsValid });
+  //         if (!answerIsValid) {
+  //           const message = ANSWERING_ERROR;
+  //           setVoiceAssistant({
+  //             ...voiceAssistant,
+  //             listen: false,
+  //             listenType: "ANSWERING",
+  //             message,
+  //             narrate: true,
+  //             selectedAnswer: "",
+  //           });
+  //           return;
+  //         }
 
-          // const message =`You have selected a, b and c. Is this correct?`
+  //         // const message =`You have selected a, b and c. Is this correct?`
 
-          const submitOptions = getAnswersLettersOptions(transcriptProcessed, voiceAssistant.answers.length);
-          assistantRef.current.onSelectAnswers(submitOptions);
-          const message = `You have selected ${getTextSplittedByCharacter(transcriptProcessed, "-")}. Is this correct?`;
-          setVoiceAssistant({
-            ...voiceAssistant,
-            listen: false,
-            listenType: "CONFIRM",
-            narrate: true,
-            message,
-            selectedAnswer: transcriptProcessed,
-          });
-          return;
-        }
-        if (voiceAssistant.listenType === "CONFIRM") {
-          console.log("assistantActions:CONFIRM");
-          if (["y"].includes(transcriptProcessed)) {
-            const submitOptions = getAnswersLettersOptions(
-              voiceAssistant.selectedAnswer,
-              voiceAssistant.answers.length
-            );
-            const correctOptionsProcessed: { choice: KnowledgeChoice; option: string }[] = voiceAssistant.answers
-              .reduce(
-                (acu: { choice: KnowledgeChoice; option: string }[], cur, idx) => [
-                  ...acu,
-                  { choice: cur, option: QUESTION_OPTIONS[idx] },
-                ],
-                []
-              )
-              .filter(cur => cur.choice.correct);
+  //         const submitOptions = getAnswersLettersOptions(transcriptProcessed, voiceAssistant.answers.length);
+  //         assistantRef.current.onSelectAnswers(submitOptions);
+  //         const message = `You have selected ${getTextSplittedByCharacter(transcriptProcessed, "-")}. Is this correct?`;
+  //         setVoiceAssistant({
+  //           ...voiceAssistant,
+  //           listen: false,
+  //           listenType: "CONFIRM",
+  //           narrate: true,
+  //           message,
+  //           selectedAnswer: transcriptProcessed,
+  //         });
+  //         return;
+  //       }
+  //       if (voiceAssistant.listenType === "CONFIRM") {
+  //         console.log("assistantActions:CONFIRM");
+  //         if (["y"].includes(transcriptProcessed)) {
+  //           const submitOptions = getAnswersLettersOptions(
+  //             voiceAssistant.selectedAnswer,
+  //             voiceAssistant.answers.length
+  //           );
+  //           const correctOptionsProcessed: { choice: KnowledgeChoice; option: string }[] = voiceAssistant.answers
+  //             .reduce(
+  //               (acu: { choice: KnowledgeChoice; option: string }[], cur, idx) => [
+  //                 ...acu,
+  //                 { choice: cur, option: QUESTION_OPTIONS[idx] },
+  //               ],
+  //               []
+  //             )
+  //             .filter(cur => cur.choice.correct);
 
-            // const correctOption = voiceAssistant.answers.filter(cur=>cur.correct)
+  //           const isCorrect = voiceAssistant.answers.reduce(
+  //             (acu, cur, idx) => acu && submitOptions[idx] === cur.correct,
+  //             true
+  //           );
+  //           const selectedAnswer: { choice: KnowledgeChoice; option: string }[] = voiceAssistant.answers.reduce(
+  //             (acu: { choice: KnowledgeChoice; option: string }[], cur, idx) => {
+  //               const answer = voiceAssistant.selectedAnswer.includes(QUESTION_OPTIONS[idx])
+  //                 ? { choice: cur, option: QUESTION_OPTIONS[idx] }
+  //                 : null;
+  //               return answer ? [...acu, answer] : acu;
+  //             },
+  //             []
+  //           );
+  //           console.log({ selectedAnswer });
+  //           const feedbackForAnswers = selectedAnswer
+  //             .map(cur => `You selected option ${cur.option}: ${cur.choice.feedback}`)
+  //             .join(". ");
+  //           const possibleAssistantMessages = isCorrect ? ASSISTANT_POSITIVE_SENTENCES : ASSISTANT_NEGATIVE_SENTENCES;
+  //           const randomMessageIndex = Math.floor(Math.random() * possibleAssistantMessages.length);
+  //           const assistantMessageBasedOnResultOfAnswer = possibleAssistantMessages[randomMessageIndex];
+  //           assistantRef.current?.onSubmitAnswer(submitOptions);
 
-            const isCorrect = voiceAssistant.answers.reduce(
-              (acu, cur, idx) => acu && submitOptions[idx] === cur.correct,
-              true
-            );
-            const selectedAnswer: { choice: KnowledgeChoice; option: string }[] = voiceAssistant.answers.reduce(
-              (acu: { choice: KnowledgeChoice; option: string }[], cur, idx) => {
-                const answer = voiceAssistant.selectedAnswer.includes(QUESTION_OPTIONS[idx])
-                  ? { choice: cur, option: QUESTION_OPTIONS[idx] }
-                  : null;
-                return answer ? [...acu, answer] : acu;
-              },
-              []
-            );
-            console.log({ selectedAnswer });
-            const feedbackForAnswers = selectedAnswer
-              .map(cur => `You selected option ${cur.option}: ${cur.choice.feedback}`)
-              .join(". ");
-            const possibleAssistantMessages = isCorrect ? ASSISTANT_POSITIVE_SENTENCES : ASSISTANT_NEGATIVE_SENTENCES;
-            const randomMessageIndex = Math.floor(Math.random() * possibleAssistantMessages.length);
-            const assistantMessageBasedOnResultOfAnswer = possibleAssistantMessages[randomMessageIndex];
-            assistantRef.current?.onSubmitAnswer(submitOptions);
+  //           const feedbackToWrongChoice = !isCorrect
+  //             ? `The correct choice${correctOptionsProcessed.length > 1 ? "s are" : " is"} ${correctOptionsProcessed
+  //                 .map(c => c.option)
+  //                 .join(" ")}`
+  //             : "";
+  //           setVoiceAssistant({
+  //             ...voiceAssistant,
+  //             listen: false,
+  //             listenType: "NEXT_ACTION",
+  //             narrate: true,
+  //             message: `${assistantMessageBasedOnResultOfAnswer} ${feedbackForAnswers} ${feedbackToWrongChoice}` ?? "",
+  //             answers: [],
+  //             selectedAnswer: "",
+  //             date: "",
+  //           });
+  //         } else {
+  //           const message = ANSWERING_ERROR;
+  //           setVoiceAssistant({ ...voiceAssistant, listen: false, listenType: "ANSWERING", narrate: true, message });
+  //         }
+  //         return;
+  //       }
 
-            const feedbackToWrongChoice = !isCorrect
-              ? `The correct choice${correctOptionsProcessed.length > 1 ? "s are" : " is"} ${correctOptionsProcessed
-                  .map(c => c.option)
-                  .join(" ")}`
-              : "";
-            setVoiceAssistant({
-              ...voiceAssistant,
-              listen: false,
-              listenType: "NEXT_ACTION",
-              narrate: true,
-              message: `${assistantMessageBasedOnResultOfAnswer} ${feedbackForAnswers} ${feedbackToWrongChoice}` ?? "",
-              answers: [],
-              selectedAnswer: "",
-              date: "",
-            });
-          } else {
-            const message = ANSWERING_ERROR;
-            setVoiceAssistant({ ...voiceAssistant, listen: false, listenType: "ANSWERING", narrate: true, message });
-          }
-          return;
-        }
+  //       if (voiceAssistant.listenType === "NEXT_ACTION") {
+  //         console.log("assistantActions:NEXT_ACTION");
+  //         if (transcriptProcessed === NEXT_ACTION) {
+  //           if (!assistantRef.current) return;
+  //           assistantRef.current.nextQuestion();
+  //           return;
+  //         }
+  //         if (transcriptProcessed === OPEN_NOTEBOOK) {
+  //           if (!assistantRef.current) return;
+  //           console.log("ACTION:Open Notebook");
+  //           const parents = assistantRef.current.getQuestionParents();
+  //           setDisplayDashboard(false);
+  //           openNodesOnNotebook(selectedNotebookId, parents);
+  //           await detectElements({ ids: parents });
+  //           for (let i = 0; i < parents.length; i++) {
+  //             const parent = parents[i];
+  //             const node: Node | null = await getNode(db, parent);
+  //             if (!node) return;
 
-        if (voiceAssistant.listenType === "NEXT_ACTION") {
-          console.log("assistantActions:NEXT_ACTION");
-          if (transcriptProcessed === NEXT_ACTION) {
-            if (!assistantRef.current) return;
-            assistantRef.current.nextQuestion();
-            return;
-          }
-          if (transcriptProcessed === OPEN_NOTEBOOK) {
-            if (!assistantRef.current) return;
-            console.log("ACTION:Open Notebook");
-            const parents = assistantRef.current.getQuestionParents();
-            setDisplayDashboard(false);
-            openNodesOnNotebook(selectedNotebookId, parents);
-            await detectElements({ ids: parents });
-            for (let i = 0; i < parents.length; i++) {
-              const parent = parents[i];
-              const node: Node | null = await getNode(db, parent);
-              if (!node) return;
+  //             const message = nodeToNarration(node);
+  //             scrollToNode(parent);
+  //             await narrateLargeTexts(message);
+  //           }
+  //           // TODO: wait for next action
+  //           console.log("execute NOTEBOOK_ACTIONS ");
+  //           setVoiceAssistant({
+  //             ...voiceAssistant,
+  //             answers: [],
+  //             date: "",
+  //             listen: false,
+  //             listenType: "NOTEBOOK_ACTIONS",
+  //             message: "",
+  //             narrate: false,
+  //             selectedAnswer: "",
+  //           });
+  //           return;
+  //         }
+  //         // No valid action was selected, try again
+  //         let message = `Sorry, I didn't get your choices. ${NEXT_ACTION_ERROR}`;
+  //         await narrateLargeTexts(message);
+  //         console.log("NEXT_ACTION");
+  //         setVoiceAssistant({ ...voiceAssistant, date: new Date().toISOString(), message: "" });
+  //         return;
+  //       }
 
-              const message = nodeToNarration(node);
-              scrollToNode(parent);
-              await narrateLargeTexts(message);
-            }
-            // TODO: wait for next action
-            console.log("execute NOTEBOOK_ACTIONS ");
-            setVoiceAssistant({
-              ...voiceAssistant,
-              answers: [],
-              date: "",
-              listen: false,
-              listenType: "NOTEBOOK_ACTIONS",
-              message: "",
-              narrate: false,
-              selectedAnswer: "",
-            });
-            return;
-          }
-          // No valid action was selected, try again
-          let message = `Sorry, I didn't get your choices. ${NEXT_ACTION_ERROR}`;
-          await narrateLargeTexts(message);
-          console.log("NEXT_ACTION");
-          setVoiceAssistant({ ...voiceAssistant, date: new Date().toISOString(), message: "" });
-          return;
-        }
+  //       if (voiceAssistant.listenType === "NOTEBOOK_ACTIONS") {
+  //         setRootQuery(voiceAssistant.tagId);
+  //         setDisplayDashboard(true);
+  //       }
+  //     };
 
-        if (voiceAssistant.listenType === "NOTEBOOK_ACTIONS") {
-          setRootQuery(voiceAssistant.tagId);
-          setDisplayDashboard(true);
-        }
-      };
+  //     recognition.onnomatch = async () => {
+  //       console.log("onnomatch");
+  //       let message = "Sorry, I didn't get your choices.";
+  //       if (voiceAssistant.listenType === "CONFIRM") {
+  //         message += CONFIRM_ERROR;
+  //       }
+  //       if (voiceAssistant.listenType === "ANSWERING") {
+  //         message += ANSWERING_ERROR;
+  //       }
+  //       if (voiceAssistant.listenType === "NEXT_ACTION") {
+  //         message += NEXT_ACTION_ERROR;
+  //       }
 
-      recognition.onnomatch = async () => {
-        console.log("onnomatch");
-        let message = "Sorry, I didn't get your choices.";
-        if (voiceAssistant.listenType === "CONFIRM") {
-          message += CONFIRM_ERROR;
-        }
-        if (voiceAssistant.listenType === "ANSWERING") {
-          message += ANSWERING_ERROR;
-        }
-        if (voiceAssistant.listenType === "NEXT_ACTION") {
-          message += NEXT_ACTION_ERROR;
-        }
+  //       await narrateLargeTexts(message);
+  //       setVoiceAssistant({
+  //         ...voiceAssistant,
+  //         date: new Date().toISOString(),
+  //         message: voiceAssistant.listenType === "NEXT_ACTION" ? "" : voiceAssistant.message,
+  //       });
+  //     };
 
-        await narrateLargeTexts(message);
-        setVoiceAssistant({
-          ...voiceAssistant,
-          date: new Date().toISOString(),
-          message: voiceAssistant.listenType === "NEXT_ACTION" ? "" : voiceAssistant.message,
-        });
-      };
+  //     recognition.onerror = async function (event: any) {
+  //       console.log("xonerror", event.error);
+  //       const message = "Sorry, I cannot detect speech, lets try again.";
+  //       await narrateLargeTexts(message);
+  //       setVoiceAssistant({
+  //         ...voiceAssistant,
+  //         date: "from-error",
+  //         message: voiceAssistant.listenType === "NEXT_ACTION" ? "" : voiceAssistant.message,
+  //       });
+  //     };
+  //   };
 
-      recognition.onerror = async function (event: any) {
-        console.log("xonerror", event.error);
-        const message = "Sorry, I cannot detect speech, lets try again.";
-        await narrateLargeTexts(message);
-        setVoiceAssistant({
-          ...voiceAssistant,
-          date: "from-error",
-          message: voiceAssistant.listenType === "NEXT_ACTION" ? "" : voiceAssistant.message,
-        });
-      };
-    };
-
-    assistantActions();
-    // prevVoiceAssistant, dont add this on dependencies, this is a ref
-  }, [
-    db,
-    openNodesOnNotebook,
-    scrollToNode,
-    selectedNotebookId,
-    voiceAssistant,
-    voiceAssistant?.answers,
-    voiceAssistant?.listenType,
-    voiceAssistant?.message,
-    voiceAssistant?.selectedAnswer,
-    voiceAssistant?.tagId,
-  ]);
+  //   assistantActions();
+  //   // prevVoiceAssistant, dont add this on dependencies, this is a ref
+  // }, [
+  //   db,
+  //   openNodesOnNotebook,
+  //   scrollToNode,
+  //   selectedNotebookId,
+  //   voiceAssistant,
+  //   voiceAssistant?.answers,
+  //   voiceAssistant?.listenType,
+  //   voiceAssistant?.message,
+  //   voiceAssistant?.selectedAnswer,
+  //   voiceAssistant?.tagId,
+  // ]);
 
   return (
     <div className="MapContainer" style={{ overflow: "hidden" }}>
@@ -6665,6 +6650,23 @@ const Notebook = ({}: NotebookProps) => {
             </Button>
           </Box>
         )}
+
+        {/* assistant */}
+        {/* {voiceAssistant && ( */}
+        <Box sx={{ position: "absolute", bottom: "50px", right: "50px", zIndex: ZINDEX["assistant"] }}>
+          <Assistant
+            voiceAssistant={voiceAssistant}
+            assistantRef={assistantRef}
+            openNodesOnNotebook={openNodesOnNotebook}
+            scrollToNode={scrollToNode}
+            selectedNotebookId={selectedNotebookId}
+            setDisplayDashboard={setDisplayDashboard}
+            setRootQuery={setRootQuery}
+            setVoiceAssistant={setVoiceAssistant}
+          />
+        </Box>
+        {/* )} */}
+
         <Box sx={{ width: "100vw", height: "100vh", overflow: "hidden" }}>
           {
             <Drawer
@@ -6965,6 +6967,7 @@ const Notebook = ({}: NotebookProps) => {
           {user && displayDashboard && (
             <DashboardWrapper
               ref={assistantRef}
+              voiceAssistant={voiceAssistant}
               setVoiceAssistant={setVoiceAssistant}
               // voiceAssistantRef={voiceAssistantRef.current}
               user={user}
@@ -6976,6 +6979,8 @@ const Notebook = ({}: NotebookProps) => {
               openNodeHandler={openNodeHandler}
               sx={{ position: "absolute", inset: "0px", zIndex: 999 }}
               root={rootQuery}
+              enabledAssistant={enabledAssistant}
+              setEnabledAssistant={setEnabledAssistant}
             />
           )}
